requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aniso8601==8.0.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coverage==5.1
cycler==0.10.0
Django==3.0.7
djangorestframework==3.11.1
flake8==3.8.3
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.9
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
multidict==4.7.6
numpy==1.18.2
peewee==3.13.2
pycodestyle==2.6.0
pydantic==1.6.1
pyflakes==2.2.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
whitenoise==5.2.0
yarl==1.5.1
zipp==3.2.0
